Refactor: adição do pacote 'bloco_notas' no Service para padronização
